+++ lib/edgehog_device/storage_usage.c
@@ -9,6 +9,8 @@
 #include "edgehog_private.h"
 #include "nvs.h"
 
+#include <time.h>
+
 #include <astarte_device_sdk/device.h>
 #include <astarte_device_sdk/result.h>
 
@@ -29,9 +31,11 @@ void publish_storage_usage(edgehog_device_handle_t edgehog_device)
             { .path = "freeBytes", .individual = astarte_individual_from_longinteger(free_space) }
         };
 
+        const int64_t timestamp = (int64_t) time(NULL);
+
         astarte_result_t res = astarte_device_stream_aggregated(edgehog_device->astarte_device,
             io_edgehog_devicemanager_StorageUsage.name, "/" NVS_PARTITION_LABEL, object_entries,
-            ARRAY_SIZE(object_entries), NULL);
+            ARRAY_SIZE(object_entries), &timestamp);
         if (res != ASTARTE_RESULT_OK) {
             EDGEHOG_LOG_ERR("Unable to send syste_status"); // NOLINT
         }