+++ CHANGELOG.md
@@ -15,7 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Add support for `io.edgehog.devicemanager.StorageUsage` interface.
 
 ### Changed
-- Updated the Astarte device SDK to `v0.6.0`.
+- Updated the Astarte device SDK to `v0.6.1`.
 
 ## [0.5.0] - 2024-05-13
 ### Added